backend/python-sidecar/requirements.txt
flask>=3.0
tensorflow>=2.16
Pillow>=10.0
numpy>=1.26
sentence-transformers>=3.0
gunicorn>=22.0
